Agregado de Hasheo de passwords (login y sign up)
