+++ HomeBanking/HomeBanking/Helpers/IPasswordHasher.cs
@@ -1,0 +1,8 @@
+﻿namespace HomeBanking.Helpers
+{
+    public interface IPasswordHasher
+    {
+        bool Verify(string passwordHashed, string inputPassword);
+        string Hash(string password);
+    }
+}
