+++ HomeBanking/HomeBanking/Repositories/ClientLoanRepository.cs
@@ -1,4 +1,5 @@
 ﻿
+using HomeBanking.Data;
 using HomeBanking.Models;
 
 namespace HomeBanking.Repositories
