+++ HomeBanking/HomeBanking/Repositories/ClientRepository.cs
@@ -1,4 +1,5 @@
-﻿using HomeBanking.Models;
+﻿using HomeBanking.Data;
+using HomeBanking.Models;
 using Microsoft.EntityFrameworkCore;
 using System.Collections.Generic;
 using System.Linq;
