+++ HomeBanking/HomeBanking/Repositories/LoanRepository.cs
@@ -1,4 +1,5 @@
-﻿using HomeBanking.Models;
+﻿using HomeBanking.Data;
+using HomeBanking.Models;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
